Refactors and budgeteer prep
Planning for budgeteer with a basic implementation in place with a function to get app budgets (getBudget in common).
The idea is for each app to only spend what is available to it.
Added a monitor flag to gangManager so we can watch stats and not take any purchasing actions.
Doomslayer interval is now a flag with the same default persisted.
selfHack now has the ability to be dynamic with port  data, this is controlled by cnccmd.
clearQueue now takes a flag of which port to clear.